Fixed errors so plugin can activate.

+++ uninstall.php
@@ -1,7 +1,12 @@
 <?php
 
+if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+  exit;
+}
+
 
 delete_option( 'floatbox' );
+delete_option( 'floatbox_version' );
 
 
 ?>
